+++ application/api/views/authorization/__init__.py
@@ -1,5 +1,5 @@
-from app.utils import format_bad_request
-from app.resource import models
+from application.utils import format_bad_request
+from application.resource import models
 
 import mongoengine
